Rename deleteModal to DeleteMetadataCard.vue
